fixed #1, int type cannot handle calculated total number of iteration when it goes very big

+++ MetaString.h
@@ -87,7 +87,7 @@ public:
 
 //    void setup(const QString & resumeStr);
 
-    int computeTotal() const;
+    double computeTotal() const;
 
 
 protected:
